[TWINFACES-230] added OptionScreen
