+++ src/entities/datalist/api/index.ts
@@ -1,3 +1,3 @@
-export * from "./datalist-api";
+export * from "./api-service";
 export * from "./types";
 export * from "./hooks";
