+++ src/entities/datalist/components/resource-link/resource-link.tsx
@@ -4,17 +4,13 @@ import { isPopulatedString } from "@/shared/libs";
 import { ResourceLink } from "@/shared/ui";
 import { ListTree } from "lucide-react";
 
-type PropsDataLink = {
+type Props = {
   data: DataList;
   disabled?: boolean;
   withTooltip?: boolean;
 };
 
-export const DatalistResourceLink = ({
-  data,
-  disabled,
-  withTooltip,
-}: PropsDataLink) => {
+export function DatalistResourceLink({ data, disabled, withTooltip }: Props) {
   const link = `/workspace/datalists/${data.id}`;
 
   return (
@@ -33,4 +29,4 @@ export const DatalistResourceLink = ({
       link={link}
     />
   );
-};
+}
