+++ src/entities/option/components/resource-link/index.ts
@@ -1,2 +1,1 @@
 export * from "./resource-link";
-export * from "./tooltip";
