+++ src/entities/option/components/resource-link/resource-link.tsx
@@ -1,22 +1,20 @@
 import { isPopulatedString } from "@/shared/libs";
 import { ResourceLink } from "@/shared/ui";
 import { Option } from "lucide-react";
-import {
-  DatalistOptionResourceTooltip,
-  DataListOptionV3,
-} from "@/entities/option";
+import { DataListOptionV3 } from "@/entities/option";
+import { DatalistOptionResourceTooltip } from "./tooltip";
 
-type PropsDataLink = {
+type Props = {
   data: DataListOptionV3;
   disabled?: boolean;
   withTooltip?: boolean;
 };
 
-export const DatalistOptionResourceLink = ({
+export function DatalistOptionResourceLink({
   data,
   disabled,
   withTooltip,
-}: PropsDataLink) => {
+}: Props) {
   const link = `/workspace/datalists/${data.dataListId}/options/${data.id}`;
 
   return (
@@ -35,4 +33,4 @@ export const DatalistOptionResourceLink = ({
       link={link}
     />
   );
-};
+}
