+++ src/screens/options/index.ts
@@ -1,1 +1,2 @@
 export * from "./options";
+export * from "./option-general";
